ocr-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
easyocr==1.7.1
Pillow==10.2.0
numpy==1.26.4
pydantic==2.6.3
python-multipart==0.0.9
